added routes to get to products & experiences pages; reorganized components/sections
